pyproject.toml
[project]
name = "graphrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "neo4j>=5.25.0",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "pandas>=2.2.3",
    "patito>=0.7.0",
    "polars>=1.9.0",
    "pyarrow>=17.0.0",
    "langchain>=0.3.3",
    "langchain-openai>=0.2.2",
    "langchain-community>=0.3.2",
    "langchain-experimental>=0.3.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
